Separate out Interpretation component
